refactor: split big file realtime recorder into smaller chunks
