+++ socket-server/internal/ws/handler.go
@@ -38,7 +38,7 @@ func (c *Client) processClientAudio() {
 
 			msgType, audio, err := c.Conn.ReadMessage()
 			if err != nil {
-				log.Printf("err read message : %d, stop this client", err)
+				log.Printf("err reading message: %v; stopping this client", err)
 				return
 			}
